requirements.txt
discord.py
python-dotenv
openai>=1.0.0
